+++ app/stackage-update.hs
@@ -6,12 +6,21 @@ main :: IO ()
 main = do
     args <- getArgs
     case args of
-        ["--help"] -> putStrLn $ "Usage: stackage-update [--verify]\n" ++
-                      "Run this command with no arguments to update your package index.\n" ++
-                      "Provide the --verify argument to verify signatures."
+        ["--help"] -> putStrLn $ unlines
+            [ "Usage: stackage-update [--verify] [--hashes]"
+            , "Run this command with no arguments to update your package index."
+            , ""
+            , "    --verify : Verify GPG signature on the repo"
+            , "    --hashes : Download from the all-cabal-hashes repo"
+            ]
         ["--summary"] -> putStrLn "Update your package index incrementally (requires git)"
-        [] -> stackageUpdate defaultStackageUpdateSettings
-        ["--verify"] -> stackageUpdate $ setVerify True defaultStackageUpdateSettings
-        xs -> putStrLn $ "stackage-update: unrecognized argument \'" ++
-              (concat $ intersperse " " xs) ++ "\'\nTry \'stackage-update --help\' for more information"
+        _ -> stackageUpdate $ foldr addArg defaultStackageUpdateSettings args
 
+addArg :: String -> StackageUpdateSettings -> StackageUpdateSettings
+addArg "--verify" = setVerify True
+addArg "--hashes" = setRemote allCabalHashes . setDirectoryName "all-cabal-hashes"
+addArg s = error $ concat
+    [ "Did not understand argument "
+    , show s
+    , ", try 'stackage-update --help' for more information"
+    ]